fix: Modal example code refactoring
