+++ src/playground/components/ModalExample/index.ts
@@ -1,0 +1,1 @@
+export { ModalExample } from './ModalExample';
